ExpressionTreeResolutionValidator
Summary:
When generating ABIs using just source, we have to make some guesses.
When we guess wrong, the failure will not show up until later in
compilation or even at runtime. To facilitate debugging, we'll check
our guesses when we do full compilation of the same source code, so
that we can provide more actionable errors.

This commit introduces the class that will do the validation, though
for the moment the only validation it does is to confirm that we parsed
the type successfully.

Test Plan: New unit test

Reviewed By: dreiss

fbshipit-source-id: 09ff00c