ValidatingTaskListener
Summary:
When generating ABIs using just source, we have to make some guesses.
When we guess wrong, the failure will not show up until later in
compilation or even at runtime. To facilitate debugging, we'll check
our guesses when we do full compilation of the same source code, so
that we can provide more actionable errors.

This commit adds the `TaskListener` that will actually run the
verification.

Test Plan:
Change the `ExpressionTreeResolutionValidatorTest` to use this
`TaskListener`.

Reviewed By: dreiss

fbshipit-source-id: 6dd858d