Plumb ABI verification setting into Javac
Summary:
When we’re constructing ABIs from source, we have to make some guesses about what some symbols mean. Those guesses will be validated when we later do a full compilation of the source.

This commit plumbs through a boolean to control whether we do that validation.

Test Plan: Actually use the boolean in the next commit

Reviewed By: dreiss

fbshipit-source-id: 95921cd