Hook up skeleton ABI validator
Summary:
When we're constructing ABIs from source, we have to
make some guesses about what some symbols mean.
Those guesses will be validated when we later do a full
compilation of the source.

This commit finally hooks up the validator to run at
compile-time. The validator doesn't actually validate
anything yet, and the feature is still off by default, so
this is just getting the pipeline in place.

Test Plan:
Set `abi_generation_mode = migrating_to_source` in the `[java]` section of `.buckconfig` and build Buck.

It should build clean, with small tracing sections visible
when the parsing and validation are running.

Reviewed By: dreiss

fbshipit-source-id: f817ef8