cxx: make `prebuilt_cxx_library()` support versioning
Summary:
This makes the `prebuilt_cxx_library()` rule support versioning
and adds support for a version sub dir, so that different prebuilds
of the same library can be selected dependening on which version
choices are selected (e.g. there may be multiple prebuilds of the
same lib against different versions of libevent, and we'll only know
when to pick the correct one after version resolution).

Test Plan: added unittest

Reviewed By: beefon

fbshipit-source-id: f8dc33a