Add support for arbitrary generated code
Summary:
This is a generic solution to support arbitrary generated code in IntelliJ project.

The basic idea: when there is a dependency on specific target we add a specific folder in `buck-out/gen/path/to/target/name_from_config` to the project. `.buckconfig` contains a map of target name to folder name. This gives us enough flexibility to adjust most of rules that generate code in our code base to be available in IntelliJ.

Another minor change is specifying `generated="true"` in generated folders. This results in showing a warning in IntelliJ when source code is modified by a user.

Test Plan:
CI

Also, run `buck project` on out code base and make sure generated code is recognized.

Reviewed By: asp2insp

fbshipit-source-id: b440f71