targets: support option to enable versioning when using action graph
Summary:
This uses the "build"-specific buck config entry which enables
versioning to gate whether versioing is used for targets command
which build the action graph.  This works well, as if versioning is
enabled for building it means `versioned_alias` rules are present
in the target graph, which cannot be transformed to actions without
versioning.

This also allows keeping versioning disabled for targets commands
which don't require an action graph, which is useful for speed
and debugging.

Test Plan: CI

Reviewed By: illicitonion

fbshipit-source-id: 8e9ee40