Log buck linker commands
Summary: The linker command is logged as any other shell command, but points to an argfile that has been since removed. This logs the entire command without the argfile, in a format where you can copy & paste to execute it manually for debugging purposes.

Test Plan: Ran buck build that uses this step.

Reviewed By: illicitonion

fbshipit-source-id: 59ae396