Fix buck's code-coverage when jar_spool_mode=direct_to_jar
Summary:
When jar_spool_mode is direct_to_jar, we don't generate a directory of
class files. But the code coverage reporting step needs such a
directory to figure out what classes are being tested.

In this diff, I just extract the jar file to that directory just
before calling the reporting step.

Test Plan:
I ran the `buck test --code-coverage` in an external repo which had the
issue and looked at the generated report and it looked fine.

In the buck repo, added a happy path test to cover the code extraction
paths.

Reviewed By: jkeljo

fbshipit-source-id: 7c7c85e