Run DX with the --no-locals flag for release builds
Summary:
For release builds we will soon add the option to skip the ProGuard stage.
In order for Redex to produce code quality code we need DX to be run
with the `--no-locals` flag for when we make release builds without ProGuard.

Test Plan: Manual testing of production apps to make sure release builds still work.

Reviewed By: dreiss

fbshipit-source-id: 21ccba2