Split SourcePathResolver into two interfaces
Summary:
This is a pure refactor, there are no behaviour changes.

SourcePathResolver resolves SourcePaths to filesystem locations.
It should only be used when constructing Steps.
It is actually used all over.

SourcePathRuleFinder resolves SourcePaths to the BuildRule which creates it.
It should really not exist, as these dependencies should be added
automatically by the parser. Until then, though, it should be used in
many fewer places than it currently is.

Both of these types are used in a lot of places they shouldn't be. This
is the first step in stopping using them where they sholdn't be used.

Test Plan: CI

Reviewed By: plamenko

fbshipit-source-id: 2be489a