Use exception mechanism instead of Optional when size limit is exceeded
Summary: Optionals are another abstraction that is heavily misused accross the Buck's codebase. This is practically a schoolbok example of when exception should be used. Not only this removes this Optional clutter all around rule key factory, but it also unwinds immediately instead of finishing iterating through all the fields only to do nothing.

Test Plan: tests

Reviewed By: illicitonion

fbshipit-source-id: 7e5412a