Refactor DependencyFileRuleKeyFactory [1]
Summary:
DefaultDependencyFileRuleKeyFactory extends InputBasedRuleKeyFactory. However, by analyzing the code, there are no grounds for actually doing that. The upcoming diffs will demonstrate that in fact there is almost nothing in common between these two, and they should not share the same hierarchy chain.
If you carefulyl look at what the implementation does, all the methods of interest do something like this:

  if (dep-file-factory) {
    do one thing
  } else if (input-based) {
    do another thing
  }

This diff, the first in the stack, just duplicates the original InputBasedRuleKeyFactory to be used specifically for DependencyFileRuleKeyFactory. In the next diff I'll remove all the bits from each factory that are not necessary for it which will clearly show that these two are mostly mutually exclusive.

Test Plan: tests

Reviewed By: illicitonion

fbshipit-source-id: e90eac2