default to UTF-8 for process executor output
Summary: DefaultProcessExecutor currently defaults to interpreting the output of invoked commands as ASCII, which garbles GHC output. UTF-8 seems like a more reasonable default. We could also make this a parameter, though that will probably touch a bunch of code.

Test Plan: I ran it locally with a Haskell project and the output displays correctly now. Happy to write some unit tests if the reviewers deem that worthwhile.

Reviewed By: illicitonion

fbshipit-source-id: 2daf3da