Disable results-cache when running with code-coverage
Summary:
In theory it might be possible to do cache the code-coverage,
but in practice we don't do it right now, so the only reasonable
behavior is to turn off results caching if buck test is called with
--code-coverage.

Test Plan:
Added a small unit test.

Also in an external repo, I did `buck test --code-coverage
//small/tar:get` multiple times and made sure the results weren't
showing up as `CACHED`.

Reviewed By: strulovich

fbshipit-source-id: dff518c