add `CxxIncludePaths` util class
Summary:
Builder for a triple-of-lists (for `-I`, `-isystem`, and `-F` paths), deduping elements in each.

When a PCH is to be built, it should use the include paths which would be used in its dependency rules' builds.  This allows for neater encapsulation of all those flag lists, as well as handy things like concatenation, converting to string lists, converting to `CxxToolFlags` objs.

Test Plan:
* Watch CI
* This is a standalone diff which just adds a class; essentially it adds dead code.  Will be used further down the stack.

Reviewed By: yiding

fbshipit-source-id: 78a8d6ab