Add 'type' arg to genrule
Summary:
In logging, all genrules are logged with type 'genrule'. This makes it
difficult to analyze specific categories of genrule in isolation (as
all the different categories end up in one cluster).

Add a 'type' parameter to the different genrules and append this to
the default getType().

Test Plan: Added test, CI.

Reviewed By: dreiss

fbshipit-source-id: f6c0128