Make rage integration tests more reliable
Summary:
The `testUpload` method was occasionally failing due to
there not being any healthy servers to upload to. This was caused
by the `/status.php` check returning 404 instead of a 200.
This is because a Handler needs to signal it handled a request by setting the
`handled` property to true, which the AbstractHandler implementations in
tests were not doing. If the test "won" the race with the first health
check from the SLB it would pass, otherwise it would fail.

Test Plan:
re-run `buck test test/com/facebook/buck/rage` 20x on my
machine. Before diff it would fail ~50% of the time, afterwards it
was successful 100% of the time.

Reviewed By: ruibm

fbshipit-source-id: e7010e4