Quick fix for rule key instability due to action graph evaluation order
Summary:
There were two different ways that transitive cxx preprocessor input is
computed. One via CxxLibrary's logic, the other via inline code in
CxxLibraryDescription::getTransitivecxxPreprocessorInput. These two pieces of
logic should return the same results, as which one is used depended on whether
CxxLibrary has been instantiated in the action graph at the time of call. This
piece of state is of course non-determinstic and global. Having 2 different
bits of code that attempts to do the same thing makes it difficult to keep in
sync. Currently, the logic inside CxxLibrary would only create an exported
header rule if there are exported headers, while the logic inline will always
create one. This leads to non-determinstic rule-key discrepencies.

In lieu of removing this duplicate code path, which is difficult due to rules other than
CxxLibrary also going through this code path, I simply change the behaviors
to match.

Test Plan:
The scenario which produced the discrepency is:

Let cxx_library A be a library with no exported deps.
Let cxx_library B be a library that depends on A.
Run buck targets --show-rulekey A#static B#static
The rulekey reported for A will be different depending on whether B has been processed first.

Run this scenario under debugger, verify that in both cases (CxxLibrary is available, or not) the rulekeys are the same.

Reviewed By: andrewjcg

fbshipit-source-id: 48c7c17