Fixes issue where partial lines are logged to machine readable log
Summary:
This was caused by a concurrency bug in MachineReadableLogger, where there was an unsafe access to the output stream.

It was also previously possible for an exception to be thrown after bytes had already been written. This code has now been moved, so that this isn't possible.

Test Plan: ci

Reviewed By: aiked

fbshipit-source-id: 82c2568