TreeBackedTypeElement.getSuperclass for tree-backed, fully-qualified, non-parameterized classes
Summary:
To support generating ABIs from source, we'll be implementing most
of the `javax.lang` interfaces using just the information provided
by the Compiler Tree API (`com.sun.source.tree`).

This commit partially implements `TreeBackedTypeElement.getSuperclass`.
It works for classes that are backed by trees in the current compilation run,
are not parameterized (or are referenced as raw types), and are specified
via fully-qualified name.

Test Plan: New unit tests

Reviewed By: dreiss

fbshipit-source-id: 48cd794