pch: fix include-path building for prefix headers
Summary:
Not necessary to do this extra appending of include paths, since here it's known that the list of include paths are the same.  Only for precompiled headers (from which the PCH contents were derived from some rule while possibly has some deps) would this be needed.  Conditional was added for this latter case.

The work being done in the prefix header case may be harmless, it's appending the same list of flags to itself, then de-duping; but the needless object building, appending, deduping, is resulting in a potentially expensive no-op.

Test Plan: Watching CI, should still work fine.  No behavioral change expected.

Reviewed By: yiding

fbshipit-source-id: cec57e4