Update ddmlib, android tools jars
Summary:
This story starts with `buck test //path/to/instrumentation_test` not working.
But only when connected to an Android emulator.

Turns out, as part of recent emulator work, the protocol was altered. Buck's
ddmlib was no longer able to get the name of the emulator.

Okay, let's just update ddmlib. Easy, right?
Wrong.

ddmlib goes hand-in-hand with all the android-tools binaries. They all depend
on each other, so if you update one, you need to update all of them.

Okay, let's do that.

Except internal APIs have changed. Of course. Let's fix that.

Except we also use internals from layoutlib (I think?) in order to merge
resources. So now we have to fix that too.

Oh, and don't forget that ddmlib now requires guava, so you have to add that to
a bunch of classpaths.

At the end of all this, we finally get instrumentation tests running again.
But my, what a hassle!

Test Plan:
Existing tests.
`buck test //path/to/instrumentation_test` works when connected to an android
emulator running Marshmallow.

Reviewed By: marcinkosiba

fbshipit-source-id: 2415cca