Add StandaloneWildcardType
Summary:
To support generating ABIs from source, we'll be implementing most
of the `javax.lang` interfaces using just the information provided
by the Compiler Tree API (`com.sun.source.tree`).

This commit adds support for creating and comparing `WildcardType`s.
It's not possible to reference such a type as the superclass of a
class (which is all we support on the `Element` side right now), so
we won't see this used for a little while.

Test Plan: New unit test

Reviewed By: asp2insp

fbshipit-source-id: dd03bb3