Better align API with its javac equivalent
Summary:
To support generating ABIs from source, we'll be implementing most
of the `javax.lang` interfaces using just the information provided
by the Compiler Tree API (`com.sun.source.tree`).

The entry point to `javac` is a `JavacTask`, which controls the
compiler and provides access to the `Elements`, `Types`, and
(indirectly) `Trees` implementations. Our `TreeResolver` performed
the latter role, but not the former. This commit adds some compiler
control logic to it, in the process renaming it `FrontendOnlyJavacTask`
to reflect its new role. Later on the tree resolution part will be
factored out into a new resolver class, which we can then pass
around to other places.

Test Plan: New unit tests

Reviewed By: dreiss

fbshipit-source-id: c96c03a