Make TreeBackedScope lazy
Summary:
To support generating ABIs from source, we'll be implementing most
of the `javax.lang` interfaces using just the information provided
by the Compiler Tree API (`com.sun.source.tree`).

This commit makes `TreeBackedScope` lazily figure out the type it
is associated with. This is not so much for efficiency as for allowing
a scope object to be created before the corresponding `Element`. With
that ability, we can provide the scope to the `Element` at construction
time, and then the `Element` can lazily resolve itself as needed.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 797d991