Add TypeParameterElements to TypeElements from outside
Summary:
To support generating ABIs from source, we'll be implementing most
of the `javax.lang` interfaces using just the information provided
by the Compiler Tree API (`com.sun.source.tree`).

This commit moves the creation logic for `TypeParameterElement`s
out of `TreeBackedTypeElement` and into `FrontendOnlyJavacTask.enterTree`.
This makes it completely impossible to be missing those elements at the
time they are needed, and will also be more in line with how methods
and fields will be created when we get there.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 46a4e6d