Preprocessors: make `WindowsPreprocessor` a subclass of `AbstractPreprocessor` as well
Summary: The stuff in the base class was repeated in WindowsPreprocessor; just extend the base class, drop duplicated methods.  Moved the trivial colorizer method (returning just an empty list) to the base class as well, to make non-colorizing the default, and overridable in subclasses (as they are in the other two instances, GccPreprocessor and ClangPreprocessor).

Test Plan: Watching CI

Reviewed By: Coneko

fbshipit-source-id: 7afd64d