Support maven classifier when download maven artifacts
Summary:
Fixes #1014
According to documentation, MavenUrlDecoder should support classifier,
but it does not. Unfortunately, since both host and a classifier can have similar
format, there is an ambiguity as to which one is optional, when one of them is not
supplied. To resolve this an extra `!` is now required in order to override a
default maven repository, which breaks backwards compability for clients using
and old way to pass a host, but should not make any difference for everyone else.
Personally, I think that having `!` is actually nice, since it draws more attention
to a fact that a non-default maven repository host is used.

Test Plan:
$ buck test //src/com/facebook/buck/file/...
also, added
  remote_file(
    name = 'groovy-groovysh-indy_remote_jar',
    url = 'mvn:org.codehaus.groovy:groovy-groovysh:jar:indy:2.4.1',
    sha1 = '1600fde728c885cc9506cb102deb1b494bd7c130',
  )
and ran
  $ buck fetch :groovy-groovysh-indy_remote_jar

Reviewed By: illicitonion

fbshipit-source-id: cbbf1ad