Run PreprocessorDelegate.checkForConflictingHeaders when constructing.
Summary:
Instead of calling this when compiling every file that use it, call it when constructing the PreprocessorDelegate object itself.

This avoids the redundant checks, makes the error report happen earlier, and gives a better error message (since the library's build target is stamped on the message, instead of the compilation rules')

Saves about a second when compiling a library of a thousand files.

Test Plan:
Profiled the run of:

  NO_BUCKD=1 ~/buck/bin/buck build somebiglibrary --deep --no-cache

Reviewed By: Coneko

fbshipit-source-id: 512efbb