Pull AbstractBuildRule#getResolver into sub-class
Summary:
This is a pure refactor.

Most BuildRules don't use a SourcePathResolver so don't need to be
passed one, but they're given them because the superclass requires it.

This will allow me to remove SourcePathResolver from being passed to
the constructors of every BuildRule incrementally, rather than needing
to remove them all in one terrifying uber-diff.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 4b8b536