Fix dep file false positives for APs
Summary:
Two scenarios where dep-file-based rulekeys give a false positive were identified:

1. A new metadata file is added to `META-INF` which an AP picks up via pattern matching on the list of files in the classpath, rather than trying to open directly by name
2. A new library dependency is added containing metadata which an AP wants to read, but no java classes are used.

In both of these cases, the library that wants to run the AP would get a dep-file-based hit, which is a false positive, since there is new metadata that the AP wishes to consume.
There are other scenarios which can trigger similar behavior that are not related to APs and their metadata, but the solution there is more involved/expensive, so we're going to
solve this problem first and come up with a more general solution in the future.

The solution provided in this diff is to explictly track the filenames of metadata objects in the dep-file-based rulekey.

Test Plan: New integration tests which explicitly cover the two scenarios above, and ensure that we don't introduce false negatives

Reviewed By: jkeljo

fbshipit-source-id: b768908