Make SymlinkTree's output path relative
Summary:
A `SymlinkTree` currently cannot be used as a target of a
`SourcePath` because its `getPathToOutput()` returns an absolute
path which breaks a check in `SourcePathResolver`:

https://github.com/facebook/buck/blob/13b9e5e06dce3133fbe92560338829b0eed73cb7/src/com/facebook/buck/rules/SourcePathResolver.java#L128

To fix that, make the output path relative.

This change tries to preserve the behavior of the `getRoot()`
function since some places rely on it being absolute (for example
C++ include paths).

Test Plan: CI.

Reviewed By: illicitonion

fbshipit-source-id: 4c0e05d