Change runtime deps to source paths.
Summary:
Runtime deps being build rules means the implementations in the rules
have to either have access to a SourcePathRuleFinder, or be passed
build rules in the constructor, thus moving the logic outside of the
rule itself. By changing them to be source paths we can avoid that.

Test Plan: CI

Reviewed By: illicitonion

fbshipit-source-id: 192b42c