Move complex setters out of RuleKeyBuilder.setSingleValue
Summary:
Right now setSingleValue calls some methods that are supposed to be overridden by the subclasses. These means that subclasses can in turn call setReflectively or what not from their implementation which cotnradicts what setSingleValue is about. This change moves this logic to setReflectively that is in charge of recursive invocations.
This should be a no-op refactor.

Test Plan: tests

Reviewed By: illicitonion

fbshipit-source-id: 69f6cad