Introduce RuleKeyHasher
Summary:
This abstraction is on a somewhat higher level than Guava Hasher in that it knows about some Buck specifics such as: `Sha1HashCode`, `ArchiveMemberPath`, ``NonHashingPath`, ...
Additionally, unlike Guava's Hasher, RuleKeyHasher does delimitting which is handy as we don't have to do it in the RuleKeyBuilder.
This will allow us to convert the current `RuleKeyLogger` to be of the same interface. This will in turn allow us to wrap the actual hasher and the logger hasher into one hasher that delegates to both, which will allow us to clean up the the RuleKeyLogger related clutter in the RuleKeyBuilder class. This will separate concerns and also make it a lot harder to make the two diverge unintentionally (which they do at the moment!).
Lastly, this will allow me to add yet another implementation, which I'm going to need for the `change propagation visualization tool`, which uses a different underlying hashing mechanism.

Right now, this should be a no-op refactor.

Test Plan: tests

Reviewed By: illicitonion

fbshipit-source-id: 8b0ca35