rust: refactor construction of action graph
Summary:
Put more logic in the Description classes and use RustCompileRule
as the general engine for invoking the compiler.

RustLibrary becomes a placeholder node so that the correct library
flavor can be
generated depending on the dependent's requirements.

RustBinary and RustTest generates a symlink tree for dynamic libraries
when doing a shared build.

Use Args for managing command-line options and their dependencies rather than
re-implementing it in an ad-hoc way.

Test Plan: CI

Reviewed By: andrewjcg

fbshipit-source-id: 017b04a