Allow duplicate resources to be banned by default
Summary:
Allowing duplicates by default, and requiring that users opt-in to
safety is more error-prone than banning by default.

This gives users the choice of not relying on unsafe behaviour by
default, as they have to explicitly opt in to it.

This will be followed by a migration flipping all callers.

Test Plan: Added tests

Reviewed By: marcinkosiba

fbshipit-source-id: d2ef7c0