Avoid buffering serialized json in compilation database
Summary:
For large compdbs the additional overhead is significant, since this causes the
entire file to be buffered in memory before being written out.

Test Plan:
Ran a compdb job that generates a 500MB compdb file. See that
max heap usage dropped from 1.7 to 1.2 GB, which is consistent with eliminating
this buffer.

Reviewed By: matbd

fbshipit-source-id: c8dec2e