TreeContextScanner
Summary:
Often when walking the AST of Java code, we want to get information
about the `Element`s that correspond to the AST node. This handy
class, modeled after the Compiler Tree API's `TreePathScanner`,
helps us keep track of that.

Test Plan: Use it, a lot, in the next commit

Reviewed By: dreiss

fbshipit-source-id: ac86ccb