Plumb boot classpath info into `ValidatingTaskListener`
Summary:
To support generating ABIs from source, we'll be implementing most
of the `javax.lang` interfaces using just the information provided
by the Compiler Tree API (`com.sun.source.tree`).

The code that works with the Compiler Tree API doesn't have direct
access to Buck itself. That code can tell the difference between
types that are being compiled from source and types that are
coming from `.class` files, but it cannot tell the difference between
the normal classpath and the boot classpath. That difference is
important, because only the boot classpath will be available when
generating ABIs.

Test Plan: Use this later

Reviewed By: dreiss

fbshipit-source-id: d54d770