Don't clean up files concurrently
Summary:
This races across multiple parallel calls to store, and can cause a file
to be attempted to be deleted more than once, which logs a nasty error.

Test Plan: I wrote a test which performs a bunch of deletes in parallel and shows that the error is logged before, and isn't after, but as the only observable effect is logspam, it didn't seem worth hooking some kind of log peeker to write a meaningful assert. So I deleted the test.

Reviewed By: aiked

fbshipit-source-id: a33d4a2