Distributed build commands now use distributed config when instantiating cache.
Summary:
Any local config (i.e. the build slave's .buckconfig) will act as an override on top of the distributed build config, and this combination will then be used to instantiate the cache.

This is necessary to ensure we have the correct repository information for the build.

Test Plan: ci

Reviewed By: ruibm

fbshipit-source-id: 8f0648a