Don't duplicate dependent resources of shared library binaries into the main bundle
Summary:
Right now when we include a dependency on a shared library (part of a `framework` bundle) that has `apple_resource` dependencies of its own, we unnecessarily copy those resources into the main bundle as well.

It turns out `AppleBuildRules.getRecursiveTargetNodeDependenciesOfTypes()` already has logic to stop recursing when we reach a `#shared` `AppleLibrary` or `CxxLibrary`.  Currently this is done for collecting link-time dependencies, but not copying resources. This diff fixes this by extending this check to resources as well.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 7320bbf