parser: don't set an empty `PYTHONPATH` when starting up py parser
Summary:
We currently set `PYTHONPATH` to the empty string when there isn't
an explicit option for user-provided python path set in the `.buckconfig`.
However, this has the side-effect of making python add the current
working directory to the search path.

This diff only sets `PYTHONPATH` if an option is specific in the
`.buckconfig`.

Test Plan: CI

Reviewed By: yiding

fbshipit-source-id: e3778e9