When using a test filter with JUnit or TestNG, check the filter before class-loading.
Summary:
When running JUnit/TestNG tests, we pass the runner a list of all possible classes
that *might* contain tests.  If no filter is provided, then we have to load each
one of those classes to check and see if there are any tests inside.  But if a
filter *is* provided, we don't necessarily have to spend time classloading.  If
the classname is completely excluded from the filter, we don't have to load that
class and check each method.

Test Plan: Added unit tests, run continuous integration tests.

Reviewed By: marcinkosiba

fbshipit-source-id: f566be2