Hack to reduce HeaderPathNormalizer memory usage
Summary:
This is something of a huge hack to reduce HeaderPathNormalizer memory usage.

HeaderPathNormalizer instances retains many Path objects for mapping dep file
entries to source paths. These paths encompass headers in all transitive
dependnecies of a particular cxx library, so they tend to get very big.

In lieu of rewriting PreprocessorDelegate/HeaderPathNormalizer to improve
sharing of data in these structures, this is a quick hack that halves the
memory consumption.

I observed that UnixPath, which is the Path implementation used on *nix
systems, will cache a string representation of a path whenever Path.toString is
called. Since we call this when filtering paths through path shortener in
HeaderPathNormalizer.Builder, this string is guaranteed to be initialized.

I add a function MorePaths.compact which rebuilds a path before inserting it
into the Map for HeaderPathNormalizer. This gets rid of the stringValue cache
for objects being stored in the map.

Needless to say this is a rather fragile hack that can be broken if the
implementation of Path changes, or if HeaderPathNormalizer is used differently
(namely, in a way that invokes toString on the keys), so a more principled
solution is still necessary.

Test Plan:
1. Run a build without cache, attach yourkit profiler.
2. While buck is compiling individual files, take a memory snapshot.
3. Look at HeaderPathNormalizers in the snapshot, confirm that stringValue
   entries are not initialized in the map keys.

Reviewed By: Coneko

fbshipit-source-id: 8ceb41c