Update manifest merger
Summary:
This switches to `ManifestMerger2` implementation and fixes https://github.com/facebook/buck/issues/299

Majority of this PR is updating aosp to `25.2.0` version of the tools.

The only interesting change is in `GenerateManifestStep.java`.

For the sake of not maintaining two versions of aosp, this PR replaces old `ManifestMerger` with `ManifestMerger2`. Having both is possible (we could take `ManifestMerger` from tools `24.x.x`, the last version that shipped with it) but if the new merger works for everyone, probably best to just go with that.
Closes https://github.com/facebook/buck/pull/1018

Reviewed By: dreiss

fbshipit-source-id: edc72b0