Don't propagate exceptions to top level when buck is dying
Summary:
These exceptions indicate that something has already gone wrong in the
build (probably a user ran ^C). Exceptions will still be logged to the
buck log, in case they're useful for debugging, but don't spam the
console or bubble up to the top level.

Test Plan: Manually made something throw one of these exceptions, saw that it logged to buck log but not to the console.

Reviewed By: aiked

fbshipit-source-id: 59f4639