Introduce concept of "interesting" inputs to dep-file rule key factory
Summary:
Right now dep-file factory has a special case that covers META-INF within archives, and this is too language and usecase specific for a rule key factory. Instead, it is better to provide a less specific mechanism to the interface and make build rules implement that correctly. Here we introduce concept of files for which existence is of interest, which is how META-INF get handled.

I decided to go with a Predicate, instead of other alternatives because:
* `Set<SourcePath> getInterestingPaths()` requires constructing a set, even if condition can be a simple test.
* `boolean isExistenceOfInterest(SourcePath)` may not be answerable by a simple test, but instead we may actually have to construct a set of interesting paths. However, in that case the set would have to be kept in the BuildRule, and possibly for much longer than necessary which would be memory inefficient.
* `Predicate<SourcePath> getExistenceOfInterestPredicate` allows us to have the best from the both worlds. For simple test, predicate can just wrap an if-condition. When constructing a set of inputs is the only way, this set can be constructed along with the predicate and kept in it so it will live only so long as the predicate itself.

Test Plan: tests (DependencyFileRuleKeyFactoryTest, AndroidLibraryAsAnnotationProcessorHostIntegrationTest)

Reviewed By: illicitonion

fbshipit-source-id: 647c198