Facilitate removing FS access from AndroidResource transformation
Summary:
Add an option to specify the inputs for `android_resource` as a map
instead of passing the top-level directory as a source path so that
we can avoid filesystem access during the target graph to action
graph transformation.

With this change, there are three possible ways resources and
assets can be treated:
  # They are specified as a PathSourcePath. This used to be the
    most common way. In this case the contents of the directory
    are still traversed during the transformation, but then the
    filtered files are placed in a symlink tree to ensure the
    filtered out files are not used.
  # They are specified as a BuildTargetSourcePath. This was not as
    common and the behavior in this case does not change. There is
    no FS access and the output of the rule is used in whole.
  # They are specified as a map from relative path to source path.
    This is a new option that behaves similarly to the first case,
    except no FS access is necessary.

Future plan is to migrate all users to cases 2 or 3 and then either
disable case 1 or remove the filesystem access from it which will
mean that no filtering will take case and rule keys might diverge
if ignored files are present.

Test Plan: * `buck test`.

Reviewed By: Coneko

fbshipit-source-id: 7af04c3