Allow tags to be added to build targets
Summary:
This diff introduces the concept of tags which can be used by any target and queried using buck query to determine if there are targets being incorrectly used. For example, if a set of targets should only be included in application X we can include the following line in any target which should only be in X;

  tags = [ 'X-only' ],

We can see if they're included in build target Y using the following query;

  buck query "attrfilter(tags,'X-only',deps(Y))"

Similarly we can use other tags to perform other queries over all or parts of the target graph to determine if certain tags are present in areas where they should not be.

Implementing tags this way ensures they aren't evaluated at build-time (ensuring there's no increase in the friction of development), and ensures that they are backwards compatible because there are no hard dependencies on their existence.

In the future we can expand on these to allow targets to specify rules relating to the tags which should, or should not, be in their dependency graph, which will allow us to detect the introduction of potentially harmful dependencies.

Test Plan: Manual testing at the moment, but I know asp2insp wants me to code some up so this may change.

Reviewed By: strulovich, illicitonion

fbshipit-source-id: cf099c5