Fix handling of prebuilt library header maps
Summary:
Currently, the way that header maps works is that headers are
symlinked in buck-out just like non-header maps and then the header map
maps from the include path to the full path relative to buck-out. Then,
when compiling we add both -Ibuck-out and
-Ibuck-out/path/to/header.hmap.

If we try to have a header map based library be a system include (which
we do for prebuilt_cxx_library), we then add '-isystem buck-out'. The
problem with this is that if a directory is added with both -I and
-isystem, the -I one is ignored. Then, *all* header map based includes
actually become -isystem includes.... oops.

This makes it so that when doing a system include, we just do '-isystem
buck-out/path/to/header/root' and don't use the header map.

Test Plan: build and run tests.

Reviewed By: andrewjcg

fbshipit-source-id: d4c611b