cancel all futures on ClosedByInterruptException
Summary:
The goal of this change is to make `ClosedByInterruptException` or wrapped `InterruptedException` to cancel
the rest of the futures so we don't have to go into the rest of them after one of them has been interrupted.

This does not change anything for the case of `keepGoing`.

Test Plan:
Manual testing using a debugger and checking what happens with the logs.
Plan:
- Make `ReflectiveRuleKeyFactory` throw a wrapped `ClosedByInterruptException` & `InterruptedException`.
- Check how `Build.java` behaves and make sure that `buildFuture.cancel(true);` gets called.
- Make sure that the exceptions continue to get propagated for other parts of Buck.

Reviewed By: Coneko

fbshipit-source-id: ef86b3f