Support running java tests from inner classes.
Summary:
Previously, if an inner class had a Test-annotated method, the test runner would not run that test, as it assumed that the list of tests to run was identical to the names of source files.  With this change, the test runner picks up the list of classes from its compiled test jar, allowing it to support running tests in inner classes.

Note that this is will potentially affect test targets in scala or kotlin, where due to the name mangling in those languages, we may not have been running tests (before this change) that will now be picked up.

Test Plan: Make sure existing tests perform exactly the same and run all continuous integration tests.  In a later change in this stack of diffs, add tests to confirm test behavior for various flavors of inner classes.

Reviewed By: asp2insp

fbshipit-source-id: 9f1b388